update config set up
